+++ test/test_image_service.py
@@ -24,14 +24,16 @@ class CloudImageServiceTestCase(ProviderTestBase):
         # the cleanup method access to the most current values
         test_instance = None
         net = None
+        subnet = None
 
         def create_img(name):
             return test_instance.create_image(name)
 
         def cleanup_img(img):
-            img.delete()
-            img.wait_for(
-                [MachineImageState.UNKNOWN, MachineImageState.ERROR])
+            if img:
+                img.delete()
+                img.wait_for(
+                    [MachineImageState.UNKNOWN, MachineImageState.ERROR])
 
         def extra_tests(img):
             # check image size
@@ -45,7 +47,6 @@ class CloudImageServiceTestCase(ProviderTestBase):
                 self.provider, instance_name)
             test_instance = helpers.get_test_instance(
                 self.provider, instance_name, subnet=subnet)
-
             sit.check_crud(self, self.provider.compute.images, MachineImage,
                            "cb_listimg", create_img, cleanup_img,
                            extra_test_func=extra_tests)
